Move roles facet out of Blacklight namespace
